fix: Rerenders on new question
Issue:
When a user had a filter selected and at the same time added a new question, the question would be added in the store as it should be, but the ui of the Questions view component would not render, hence the ui would still show the filtered questions(previous UI state).

Solution in this commit.
A useEffect hook was added in order to "force" a rerender by passing the questions array from the redux store as a dependency.